Add files via upload
The goal of the project was to identify playing cards correctly and efficiently. 

Two methods were implemented and tested, first was using the whole card as templates, and second using the number and the suite of the card and parsing them separately for a 2 step template matching. The latter produced much faster results.

The steps involved in the development of the project are : image preprocessing, edge detection to extract the boundaries of the card, and template matching which entails scaling and rotation for a perfect match to the template.